chatbot-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
qdrant-client
langchain-google-genai
langchain
langchain-community
firebase-admin